// src/routes/graphql/types/loaders-enum.ts
export enum Loaders {
  USER = 'userLoader',
  POST_BY_ID = 'postByIdLoader',
  POST_BY_AUTHOR_ID = 'postByAuthorIdLoader',
  PROFILE_BY_ID = 'profileByIdLoader',
  PROFILE_BY_USER_ID = 'profileByUserIdLoader',
  MEMBER_TYPE = 'memberTypeLoader',
  SUBSCRIBED_TO_USER = 'subscribedToUserLoader',
  USER_SUBSCRIBED_TO = 'userSubscribedToLoader',
}